feat(lib): use`QueryWizardBloc` bloc
