+++ lib/src/query_wizard_bloc_observer.dart
@@ -1,6 +1,6 @@
 import 'package:bloc/bloc.dart';
 
-class SimpleBlocObserver extends BlocObserver {
+class QueryWizardBlocObserver extends BlocObserver {
   @override
   void onEvent(Bloc bloc, Object event) {
     super.onEvent(bloc, event);
