+++ lib/src/repositories/query_wizard_api_client.dart
@@ -3,21 +3,21 @@ import 'package:query_wizard/src/models/models.dart';
 class QueryWizardApiClient {
   Future<List<DbElement>> getSources() async {
     List<DbElement> columns1 = [
-      DbElement(name: 'Table 1', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 2', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 3', nodeType: DbNodeType.table),
+      DbElement(name: 'Field 1', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 2', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 3', nodeType: DbNodeType.column),
     ];
 
     List<DbElement> columns2 = [
-      DbElement(name: 'Table 1', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 2', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 3', nodeType: DbNodeType.table),
+      DbElement(name: 'Field 1', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 2', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 3', nodeType: DbNodeType.column),
     ];
 
     List<DbElement> columns3 = [
-      DbElement(name: 'Table 1', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 2', nodeType: DbNodeType.table),
-      DbElement(name: 'Table 3', nodeType: DbNodeType.table),
+      DbElement(name: 'Field 1', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 2', nodeType: DbNodeType.column),
+      DbElement(name: 'Field 3', nodeType: DbNodeType.column),
     ];
 
     return [
